refactor: move some logic from the tab-main of charts settings to aggregation builder #78949
