+++ libs/shared/src/lib/components/widgets/chart-settings/tab-main/tab-main.component.ts
@@ -13,7 +13,6 @@ import {
   ReferenceDataQueryResponse,
 } from '../../../../models/reference-data.model';
 import { AggregationBuilderService } from '../../../../services/aggregation-builder/aggregation-builder.service';
-import { QueryBuilderService } from '../../../../services/query-builder/query-builder.service';
 import { AggregationService } from '../../../../services/aggregation/aggregation.service';
 import { get } from 'lodash';
 import { UnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
@@ -63,14 +62,12 @@ export class TabMainComponent extends UnsubscribeComponent implements OnInit {
    * @param apollo Apollo service
    * @param dialog Dialog service
    * @param aggregationBuilder Shared aggregation builder service
-   * @param queryBuilder Shared query builder service
    * @param aggregationService Shared aggregation service
    */
   constructor(
     private apollo: Apollo,
     private dialog: Dialog,
     private aggregationBuilder: AggregationBuilderService,
-    private queryBuilder: QueryBuilderService,
     private aggregationService: AggregationService
   ) {
     super();
@@ -127,7 +124,12 @@ export class TabMainComponent extends UnsubscribeComponent implements OnInit {
         this.resource = data.resource;
         if (aggregationId && this.resource.aggregations?.edges[0]) {
           this.aggregation = this.resource.aggregations.edges[0].node;
-          this.setAvailableSeriesFields();
+          this.availableSeriesFields =
+            this.aggregationBuilder.getAvailableSeriesFields(this.aggregation, {
+              resource: this.resource,
+            });
+        } else {
+          this.availableSeriesFields = [];
         }
       });
   }
@@ -152,57 +154,14 @@ export class TabMainComponent extends UnsubscribeComponent implements OnInit {
         this.referenceData = data.referenceData;
         if (aggregationId && this.referenceData.aggregations?.edges[0]) {
           this.aggregation = this.referenceData.aggregations.edges[0].node;
-          this.setAvailableSeriesFields();
+          this.availableSeriesFields =
+            this.aggregationBuilder.getAvailableSeriesFields(this.aggregation, {
+              referenceData: this.referenceData,
+            });
+        } else {
+          this.availableSeriesFields = [];
         }
       });
-  }
-
-  /**
-   * Set available series fields, from resource fields and aggregation definition.
-   */
-  private setAvailableSeriesFields(): void {
-    if (this.aggregation) {
-      const queryName = this.resource
-        ? this.resource.queryName
-        : this.referenceData?.graphQLTypeName;
-      const fields = this.queryBuilder
-        .getFields(queryName as string)
-        .filter(
-          (field: any) =>
-            !(
-              field.name.includes('_id') &&
-              (field.type.name === 'ID' ||
-                (field.type.kind === 'LIST' && field.type.ofType.name === 'ID'))
-            )
-        );
-      const selectedFields = this.aggregation.sourceFields
-        .map((x: string) => {
-          const field = fields.find((y) => x === y.name);
-          if (!field) return null;
-          if (field.type.kind !== 'SCALAR') {
-            Object.assign(field, {
-              fields: this.queryBuilder.deconfineFields(
-                field.type,
-                new Set()
-                  .add(
-                    this.resource
-                      ? this.resource.name
-                      : this.referenceData?.name
-                  )
-                  .add(field.type.ofType?.name)
-              ),
-            });
-          }
-          return field;
-        })
-        .filter((x: any) => x !== null);
-      this.availableSeriesFields = this.aggregationBuilder.fieldsAfter(
-        selectedFields,
-        this.aggregation?.pipeline
-      );
-    } else {
-      this.availableSeriesFields = [];
-    }
   }
 
   /**
@@ -228,7 +187,15 @@ export class TabMainComponent extends UnsubscribeComponent implements OnInit {
       if (value) {
         this.formGroup.get('chart.aggregationId')?.setValue(value.id);
         this.aggregation = value;
-        this.setAvailableSeriesFields();
+        if (this.aggregation) {
+          this.availableSeriesFields =
+            this.aggregationBuilder.getAvailableSeriesFields(this.aggregation, {
+              referenceData: this.referenceData,
+              resource: this.resource,
+            });
+        } else {
+          this.availableSeriesFields = [];
+        }
       }
     });
   }
